Delete helper PublicStateManagerIT
